gstreamer-vaapi: Autospec creation for update from version 1.18.2 to version 1.18.3
commit 5890079ecbf2f9c76a5ded398cda8fd2ac822d21
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Wed Jan 13 21:13:25 2021 +0000

    Release 1.18.3

commit 0b88e2a82762f55456a5a113e7039f1de0547802
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Sun Dec 6 23:57:17 2020 +0000

    Back to development

The latest bug-fix release in the 1.18 series is 1.18.3 and was released
on 13 January 2021.

See https://gstreamer.freedesktop.org/releases/1.18/ for the latest
version of this document.

Last updated: Wednesday 13 January 2021, 20:00 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)